+++ src/redux/store.js
@@ -1,14 +1,13 @@
 import { createStore } from "redux";
-const counterReducer = (state = { counter: 0 }, action) => {
+const counterReducer = (state = { td: [] }, action) => {
   if (action.type === "increment") {
     return { counter: state.counter + 1 };
   }
-  if (action.type === "decrement") {
-    return { counter: state.counter - 1 };
+  if(action.type === "addData"){
+    return { td: [...state.td, action.value]}
   }
-
-  if (action.type === "plusTen"){
-      return {counter: state.counter + action.value}
+  if(action.type === "removeLastData"){
+      return {td: [...(state.td.pop())]}
   }
   return state;
 };
